updated with new animate diff code
